refactor: add week hook
